HW2/requirements.txt
pandas==2.2.2 
tqdm==4.66.5 
Pillow==10.3.0 
opencv-python==4.9.0.80
numpy==1.26.4
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0
ultralytics==8.3.220